containers/fargate/requirements.txt
selenium
boto3
botocore
pandas
psycopg2-binary